refactor(maven): share coordinate and URL helpers
